ECP code cleanup
git-svn-id: https://segomo.elka.pw.edu.pl/svn/mrrocpp/base/trunk@2892 4c39dba3-d726-0410-8099-c98d187771e6

+++ src/ecp/irp6_on_track/ecp_g_trg.cc
@@ -360,7 +360,7 @@ try{
     sr_ecp_msg.message (e.error_class, e.error_no);
 }catch (...){
     // Wylapywanie niezdefiniowanych bledow.
-    sr_ecp_msg.message (lib::NON_FATAL_ERROR, (uint64_t) ECP_UNIDENTIFIED_ERROR);
+    sr_ecp_msg.message (lib::NON_FATAL_ERROR, ECP_UNIDENTIFIED_ERROR);
     } // end: catch
 } // end load_trajectory
